+++ packages/contracts-communication/test/InterchainDB.Src.t.sol
@@ -23,14 +23,14 @@ import {Test} from "forge-std/Test.sol";
 /// @notice Unit tests for InterchainDB interactions on the source chain
 /// Note: we inherit from interface with the events to avoid their copy-pasting.
 contract InterchainDBSourceTest is Test, InterchainDBEvents {
-    uint256 public constant SRC_CHAIN_ID = 1337;
-    uint256 public constant DST_CHAIN_ID = 7331;
+    uint64 public constant SRC_CHAIN_ID = 1337;
+    uint64 public constant DST_CHAIN_ID = 7331;
 
     uint16 public constant DB_VERSION = 1;
 
-    uint256 public constant INITIAL_WRITER_F = 1;
-    uint256 public constant INITIAL_WRITER_S = 2;
-    uint256 public constant INITIAL_DB_NONCE = INITIAL_WRITER_F + INITIAL_WRITER_S;
+    uint64 public constant INITIAL_WRITER_F = 1;
+    uint64 public constant INITIAL_WRITER_S = 2;
+    uint64 public constant INITIAL_DB_NONCE = INITIAL_WRITER_F + INITIAL_WRITER_S;
 
     uint256 public constant MODULE_A_FEE = 100;
     uint256 public constant MODULE_B_FEE = 200;
@@ -68,28 +68,28 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function initialWrites() internal {
-        for (uint256 i = 0; i < INITIAL_WRITER_F; ++i) {
+        for (uint64 i = 0; i < INITIAL_WRITER_F; ++i) {
             InterchainEntry memory entry = getMockEntry(i, writerF);
             writeEntry(writerF, entry.dataHash);
             initialEntries.push(entry);
         }
-        for (uint256 i = 0; i < INITIAL_WRITER_S; ++i) {
+        for (uint64 i = 0; i < INITIAL_WRITER_S; ++i) {
             InterchainEntry memory entry = getMockEntry(INITIAL_WRITER_F + i, writerS);
             writeEntry(writerS, entry.dataHash);
             initialEntries.push(entry);
         }
     }
 
-    function getInitialEntry(uint256 dbNonce) internal view returns (InterchainEntry memory) {
+    function getInitialEntry(uint64 dbNonce) internal view returns (InterchainEntry memory) {
         require(dbNonce < initialEntries.length, "dbNonce out of range");
         return initialEntries[dbNonce];
     }
 
-    function getMockDataHash(address writer, uint256 nonce) internal pure returns (bytes32) {
+    function getMockDataHash(address writer, uint64 nonce) internal pure returns (bytes32) {
         return keccak256(abi.encode(writer, nonce));
     }
 
-    function getMockEntry(uint256 dbNonce, address writer) internal pure returns (InterchainEntry memory entry) {
+    function getMockEntry(uint64 dbNonce, address writer) internal pure returns (InterchainEntry memory entry) {
         return InterchainEntry({
             srcChainId: SRC_CHAIN_ID,
             dbNonce: dbNonce,
@@ -123,19 +123,12 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         vm.mockCall(address(module), callData, returnData);
     }
 
-    function writeEntry(address writer, bytes32 dataHash) internal returns (uint256 dbNonce) {
+    function writeEntry(address writer, bytes32 dataHash) internal returns (uint64 dbNonce) {
         vm.prank(writer);
         (dbNonce,) = icDB.writeEntry(dataHash);
     }
 
-    function requestVerification(
-        address caller,
-        uint256 msgValue,
-        uint256 dbNonce,
-        address[] memory modules
-    )
-        internal
-    {
+    function requestVerification(address caller, uint256 msgValue, uint64 dbNonce, address[] memory modules) internal {
         deal(caller, msgValue);
         vm.prank(caller);
         icDB.requestBatchVerification{value: msgValue}(DST_CHAIN_ID, dbNonce, modules);
@@ -148,7 +141,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         address[] memory modules
     )
         internal
-        returns (uint256 dbNonce)
+        returns (uint64 dbNonce)
     {
         deal(writer, msgValue);
         vm.prank(writer);
@@ -162,21 +155,34 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         assertEq(entryValue, expectedValue, "!entryValue");
     }
 
+    function expectEventInterchainEntryWritten(InterchainEntry memory entry) internal {
+        vm.expectEmit(address(icDB));
+        emit InterchainEntryWritten({
+            dbNonce: entry.dbNonce,
+            entryIndex: entry.entryIndex,
+            srcWriter: entry.srcWriter,
+            dataHash: entry.dataHash
+        });
+        // In the V1 of InterchainDB, the batch is finalized immediately after writing an entry
+        vm.expectEmit(address(icDB));
+        emit InterchainBatchFinalized({dbNonce: entry.dbNonce, batchRoot: entry.entryValue()});
+    }
+
     function expectEventVerificationRequested(InterchainEntry memory entry, address[] memory srcModules) internal {
         bytes32 batchRoot = InterchainEntryLib.entryValue(entry);
         vm.expectEmit(address(icDB));
         emit InterchainBatchVerificationRequested(DST_CHAIN_ID, entry.dbNonce, batchRoot, srcModules);
     }
 
-    function expectRevertBatchDoesNotExist(uint256 dbNonce) internal {
+    function expectRevertBatchDoesNotExist(uint64 dbNonce) internal {
         vm.expectRevert(abi.encodeWithSelector(IInterchainDB.InterchainDB__BatchDoesNotExist.selector, dbNonce));
     }
 
-    function expectRevertBatchNotFinalized(uint256 dbNonce) internal {
+    function expectRevertBatchNotFinalized(uint64 dbNonce) internal {
         vm.expectRevert(abi.encodeWithSelector(IInterchainDB.InterchainDB__BatchNotFinalized.selector, dbNonce));
     }
 
-    function expectRevertEntryIndexOutOfRange(uint256 dbNonce, uint64 entryIndex, uint64 batchSize) internal {
+    function expectRevertEntryIndexOutOfRange(uint64 dbNonce, uint64 entryIndex, uint64 batchSize) internal {
         vm.expectRevert(
             abi.encodeWithSelector(
                 IInterchainDB.InterchainDB__EntryIndexOutOfRange.selector, dbNonce, entryIndex, batchSize
@@ -190,7 +196,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         );
     }
 
-    function expectRevertInvalidEntryRange(uint256 dbNonce, uint64 start, uint64 end) internal {
+    function expectRevertInvalidEntryRange(uint64 dbNonce, uint64 start, uint64 end) internal {
         vm.expectRevert(
             abi.encodeWithSelector(IInterchainDB.InterchainDB__InvalidEntryRange.selector, dbNonce, start, end)
         );
@@ -200,7 +206,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         vm.expectRevert(IInterchainDB.InterchainDB__NoModulesSpecified.selector);
     }
 
-    function expectRevertSameChainId(uint256 chainId) internal {
+    function expectRevertSameChainId(uint64 chainId) internal {
         vm.expectRevert(abi.encodeWithSelector(IInterchainDB.InterchainDB__SameChainId.selector, chainId));
     }
 
@@ -211,7 +217,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_setup_getEntryValue() public {
-        for (uint256 i = 0; i < INITIAL_DB_NONCE; ++i) {
+        for (uint64 i = 0; i < INITIAL_DB_NONCE; ++i) {
             assertCorrectValue(icDB.getEntryValue(i, 0), getInitialEntry(i));
         }
     }
@@ -219,10 +225,9 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     // ══════════════════════════════════════════ TESTS: WRITING AN ENTRY ══════════════════════════════════════════════
 
     function test_writeEntry_writerF_emitsEvent() public {
-        bytes32 dataHash = getMockDataHash(writerF, INITIAL_DB_NONCE);
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerF), dataHash);
-        writeEntry(writerF, dataHash);
+        InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerF);
+        expectEventInterchainEntryWritten(entry);
+        writeEntry(writerF, entry.dataHash);
     }
 
     function test_writeEntry_writerF_increasesDBNonce() public {
@@ -233,7 +238,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntry_writerF_returnsCorrectNonce() public {
         bytes32 dataHash = getMockDataHash(writerF, INITIAL_DB_NONCE);
-        uint256 nonce = writeEntry(writerF, dataHash);
+        uint64 nonce = writeEntry(writerF, dataHash);
         assertEq(nonce, INITIAL_DB_NONCE);
     }
 
@@ -244,10 +249,9 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_writeEntry_writerS_emitsEvent() public {
-        bytes32 dataHash = getMockDataHash(writerS, INITIAL_DB_NONCE);
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerS), dataHash);
-        writeEntry(writerS, dataHash);
+        InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerS);
+        expectEventInterchainEntryWritten(entry);
+        writeEntry(writerS, entry.dataHash);
     }
 
     function test_writeEntry_writerS_increasesDBNonce() public {
@@ -258,7 +262,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntry_writerS_returnsCorrectNonce() public {
         bytes32 dataHash = getMockDataHash(writerS, INITIAL_DB_NONCE);
-        uint256 nonce = writeEntry(writerS, dataHash);
+        uint64 nonce = writeEntry(writerS, dataHash);
         assertEq(nonce, INITIAL_DB_NONCE);
     }
 
@@ -271,13 +275,13 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     // ═══════════════════════════════════════ TESTS: REQUESTING VALIDATION ════════════════════════════════════════════
 
     function test_requestVerification_writerF_oneModule_emitsEvent() public {
-        uint256 dbNonce = 0;
+        uint64 dbNonce = 0;
         expectEventVerificationRequested(getInitialEntry(dbNonce), oneModule);
         requestVerification(requestCaller, MODULE_A_FEE, dbNonce, oneModule);
     }
 
     function test_requestVerification_writerF_oneModule_callsModule() public {
-        uint256 dbNonce = 0;
+        uint64 dbNonce = 0;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE, getModuleCalldata(entry));
@@ -285,7 +289,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_requestVerification_writerF_oneModule_higherFee() public {
-        uint256 dbNonce = 0;
+        uint64 dbNonce = 0;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
@@ -294,13 +298,13 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_requestVerification_writerF_twoModules_emitsEvent() public {
-        uint256 dbNonce = 0;
+        uint64 dbNonce = 0;
         expectEventVerificationRequested(getInitialEntry(dbNonce), twoModules);
         requestVerification(requestCaller, MODULE_A_FEE + MODULE_B_FEE, dbNonce, twoModules);
     }
 
     function test_requestVerification_writerF_twoModules_callsModules() public {
-        uint256 dbNonce = 0;
+        uint64 dbNonce = 0;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE, getModuleCalldata(entry));
@@ -310,7 +314,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_requestVerification_writerF_twoModules_higherFee() public {
         // Overpaid fees should be directed to the first module
-        uint256 dbNonce = 0;
+        uint64 dbNonce = 0;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
@@ -320,13 +324,13 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_requestVerification_writerS_oneModule_emitsEvent() public {
-        uint256 dbNonce = 2;
+        uint64 dbNonce = 2;
         expectEventVerificationRequested(getInitialEntry(dbNonce), oneModule);
         requestVerification(requestCaller, MODULE_A_FEE, dbNonce, oneModule);
     }
 
     function test_requestVerification_writerS_oneModule_callsModule() public {
-        uint256 dbNonce = 2;
+        uint64 dbNonce = 2;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE, getModuleCalldata(entry));
@@ -334,7 +338,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_requestVerification_writerS_oneModule_higherFee() public {
-        uint256 dbNonce = 2;
+        uint64 dbNonce = 2;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
@@ -343,13 +347,13 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_requestVerification_writerS_twoModules_emitsEvent() public {
-        uint256 dbNonce = 2;
+        uint64 dbNonce = 2;
         expectEventVerificationRequested(getInitialEntry(dbNonce), twoModules);
         requestVerification(requestCaller, MODULE_A_FEE + MODULE_B_FEE, dbNonce, twoModules);
     }
 
     function test_requestVerification_writerS_twoModules_callsModules() public {
-        uint256 dbNonce = 2;
+        uint64 dbNonce = 2;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE, getModuleCalldata(entry));
@@ -359,7 +363,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_requestVerification_writerS_twoModules_higherFee() public {
         // Overpaid fees should be directed to the first module
-        uint256 dbNonce = 2;
+        uint64 dbNonce = 2;
         InterchainEntry memory entry = getInitialEntry(dbNonce);
         // expectCall(address callee, uint256 msgValue, bytes calldata data)
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
@@ -378,10 +382,8 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     function test_requestVerification_revert_batchNotFinalized_oneModule_hugeNonce() public {
         expectRevertBatchNotFinalized(2 ** 32);
         requestVerification(requestCaller, MODULE_A_FEE, 2 ** 32, oneModule);
-        expectRevertBatchNotFinalized(2 ** 64);
-        requestVerification(requestCaller, MODULE_A_FEE, 2 ** 64, oneModule);
-        expectRevertBatchNotFinalized(type(uint256).max);
-        requestVerification(requestCaller, MODULE_A_FEE, type(uint256).max, oneModule);
+        expectRevertBatchNotFinalized(type(uint64).max);
+        requestVerification(requestCaller, MODULE_A_FEE, type(uint64).max, oneModule);
     }
 
     function test_requestVerification_revert_batchNotFinalized_twoModules_nextNonce() public {
@@ -392,10 +394,8 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     function test_requestVerification_revert_batchNotFinalized_twoModules_hugeNonce() public {
         expectRevertBatchNotFinalized(2 ** 32);
         requestVerification(requestCaller, MODULE_A_FEE + MODULE_B_FEE, 2 ** 32, twoModules);
-        expectRevertBatchNotFinalized(2 ** 64);
-        requestVerification(requestCaller, MODULE_A_FEE + MODULE_B_FEE, 2 ** 64, twoModules);
-        expectRevertBatchNotFinalized(type(uint256).max);
-        requestVerification(requestCaller, MODULE_A_FEE + MODULE_B_FEE, type(uint256).max, twoModules);
+        expectRevertBatchNotFinalized(type(uint64).max);
+        requestVerification(requestCaller, MODULE_A_FEE + MODULE_B_FEE, type(uint64).max, twoModules);
     }
 
     function test_requestVerification_revert_incorrectFeeAmount_oneModule_underpaid() public {
@@ -431,8 +431,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerF_oneModule_emitsEvents() public {
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerF);
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerF), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         expectEventVerificationRequested(entry, oneModule);
         writeEntryWithVerification(MODULE_A_FEE, writerF, entry.dataHash, oneModule);
     }
@@ -445,7 +444,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerF_oneModule_returnsCorrectNonce() public {
         bytes32 dataHash = getMockDataHash(writerF, INITIAL_DB_NONCE);
-        uint256 nonce = writeEntryWithVerification(MODULE_A_FEE, writerF, dataHash, oneModule);
+        uint64 nonce = writeEntryWithVerification(MODULE_A_FEE, writerF, dataHash, oneModule);
         assertEq(nonce, INITIAL_DB_NONCE);
     }
 
@@ -459,8 +458,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         bytes32 dataHash = getMockDataHash(writerF, INITIAL_DB_NONCE);
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerF);
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerF), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         writeEntryWithVerification(MODULE_A_FEE * 2, writerF, dataHash, oneModule);
         assertEq(icDB.getDBNonce(), INITIAL_DB_NONCE + 1);
         assertCorrectValue(icDB.getEntryValue(INITIAL_DB_NONCE, 0), entry);
@@ -475,8 +473,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerF_twoModules_emitsEvents() public {
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerF);
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerF), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         expectEventVerificationRequested(entry, twoModules);
         writeEntryWithVerification(MODULE_A_FEE + MODULE_B_FEE, writerF, entry.dataHash, twoModules);
     }
@@ -489,7 +486,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerF_twoModules_returnsCorrectNonce() public {
         bytes32 dataHash = getMockDataHash(writerF, INITIAL_DB_NONCE);
-        uint256 nonce = writeEntryWithVerification(MODULE_A_FEE + MODULE_B_FEE, writerF, dataHash, twoModules);
+        uint64 nonce = writeEntryWithVerification(MODULE_A_FEE + MODULE_B_FEE, writerF, dataHash, twoModules);
         assertEq(nonce, INITIAL_DB_NONCE);
     }
 
@@ -504,8 +501,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerF);
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
         vm.expectCall(address(moduleB), MODULE_B_FEE, getModuleCalldata(entry));
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerF), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         writeEntryWithVerification(MODULE_A_FEE * 2 + MODULE_B_FEE, writerF, dataHash, twoModules);
         assertEq(icDB.getDBNonce(), INITIAL_DB_NONCE + 1);
         assertCorrectValue(icDB.getEntryValue(INITIAL_DB_NONCE, 0), entry);
@@ -519,8 +515,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerS_oneModule_emitsEvents() public {
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerS);
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerS), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         expectEventVerificationRequested(entry, oneModule);
         writeEntryWithVerification(MODULE_A_FEE, writerS, entry.dataHash, oneModule);
     }
@@ -533,7 +528,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerS_oneModule_returnsCorrectNonce() public {
         bytes32 dataHash = getMockDataHash(writerS, INITIAL_DB_NONCE);
-        uint256 nonce = writeEntryWithVerification(MODULE_A_FEE, writerS, dataHash, oneModule);
+        uint64 nonce = writeEntryWithVerification(MODULE_A_FEE, writerS, dataHash, oneModule);
         assertEq(nonce, INITIAL_DB_NONCE);
     }
 
@@ -547,8 +542,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         bytes32 dataHash = getMockDataHash(writerS, INITIAL_DB_NONCE);
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerS);
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerS), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         writeEntryWithVerification(MODULE_A_FEE * 2, writerS, dataHash, oneModule);
         assertEq(icDB.getDBNonce(), INITIAL_DB_NONCE + 1);
         assertCorrectValue(icDB.getEntryValue(INITIAL_DB_NONCE, 0), entry);
@@ -563,8 +557,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerS_twoModules_emitsEvents() public {
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerS);
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerS), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         expectEventVerificationRequested(entry, twoModules);
         writeEntryWithVerification(MODULE_A_FEE + MODULE_B_FEE, writerS, entry.dataHash, twoModules);
     }
@@ -577,7 +570,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
 
     function test_writeEntryWithVerification_writerS_twoModules_returnsCorrectNonce() public {
         bytes32 dataHash = getMockDataHash(writerS, INITIAL_DB_NONCE);
-        uint256 nonce = writeEntryWithVerification(MODULE_A_FEE + MODULE_B_FEE, writerS, dataHash, twoModules);
+        uint64 nonce = writeEntryWithVerification(MODULE_A_FEE + MODULE_B_FEE, writerS, dataHash, twoModules);
         assertEq(nonce, INITIAL_DB_NONCE);
     }
 
@@ -592,8 +585,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
         InterchainEntry memory entry = getMockEntry(INITIAL_DB_NONCE, writerS);
         vm.expectCall(address(moduleA), MODULE_A_FEE * 2, getModuleCalldata(entry));
         vm.expectCall(address(moduleB), MODULE_B_FEE, getModuleCalldata(entry));
-        vm.expectEmit(address(icDB));
-        emit InterchainEntryWritten(SRC_CHAIN_ID, INITIAL_DB_NONCE, addressToBytes32(writerS), entry.dataHash);
+        expectEventInterchainEntryWritten(entry);
         writeEntryWithVerification(MODULE_A_FEE * 2 + MODULE_B_FEE, writerS, dataHash, twoModules);
         assertEq(icDB.getDBNonce(), INITIAL_DB_NONCE + 1);
         assertCorrectValue(icDB.getEntryValue(INITIAL_DB_NONCE, 0), entry);
@@ -659,7 +651,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_getBatchLeafs() public {
-        for (uint256 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
+        for (uint64 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
             bytes32[] memory leafs = icDB.getBatchLeafs(nonce);
             assertEq(leafs.length, 1, "!leafs.length");
             checkBatchRoot(leafs[0], getInitialEntry(nonce));
@@ -677,7 +669,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_getBatchLeafsPaginated() public {
-        for (uint256 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
+        for (uint64 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
             bytes32[] memory leafs = icDB.getBatchLeafsPaginated(nonce, 0, 1);
             assertEq(leafs.length, 1, "!leafs.length");
             checkBatchRoot(leafs[0], getInitialEntry(nonce));
@@ -706,7 +698,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_getBatchSize_finalized() public {
-        for (uint256 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
+        for (uint64 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
             assertEq(icDB.getBatchSize(nonce), 1, "!batchSize");
         }
     }
@@ -721,7 +713,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_getBatch() public {
-        for (uint256 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
+        for (uint64 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
             InterchainBatch memory batch = icDB.getBatch(nonce);
             checkBatch(batch, getInitialEntry(nonce));
         }
@@ -738,7 +730,7 @@ contract InterchainDBSourceTest is Test, InterchainDBEvents {
     }
 
     function test_getEntryValue() public {
-        for (uint256 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
+        for (uint64 nonce = 0; nonce < INITIAL_DB_NONCE; ++nonce) {
             InterchainEntry memory expectedEntry = getInitialEntry(nonce);
             assertCorrectValue(icDB.getEntryValue(nonce, 0), expectedEntry);
         }